Cherry-pick commit 8e28be457cf120f3d8d7b2867535483cfaa4949f
MB-27243: log that we've received a request to delete a bucket

When deleting buckets we don't log prior to actually attempting the
deletion. It's a significant enough operation that adding
a debug log that we've received a request and are about to delete
seems reasonable. And it will make it simpler to know whether or
not deletion is hanging or if the request wasn't actually received.

Change-Id: I2bee816626a6f097904fd44134c004f999fa6aeb
Reviewed-on: http://review.couchbase.org/86884
Reviewed-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Blow <michael.blow@couchbase.com>
Tested-by: Michael Blow <michael.blow@couchbase.com>